Test suite for dso-token mutlitests separated to its own module
